⚙️ Move UI components to ui-library folder
