+++ client/src/layout/Library/LibraryHeader/HeaderTitle.jsx
@@ -1,6 +1,6 @@
 import styles from './HeaderTitle.module.scss';
 import { RiAddLargeFill, RiArchiveStackLine } from 'react-icons/ri';
-import ButtonIcon from '../../../components/ButtonIcon/ButtonIcon.jsx';
+import ButtonIcon from '../../../ui-library/ButtonIcon/ButtonIcon.jsx';
 
 const HeaderTitle = ({ isCollapsed, onCollapse }) => {
   // INFO: Collapse hides "Your Library" text and "+" buttons
