+++ client/src/ui-library/Avatar/Avatar.module.scss
